Issue #1138: Moved acquire permissions
