+++ resilience4j-ratelimiter/src/main/java/io/github/resilience4j/ratelimiter/internal/BaseAtomicLimiter.java
@@ -20,7 +20,6 @@ package io.github.resilience4j.ratelimiter.internal;
 
 import io.github.resilience4j.ratelimiter.RateLimiter;
 import io.github.resilience4j.ratelimiter.RateLimiterConfig;
-import io.github.resilience4j.ratelimiter.RefillRateLimiterConfig;
 import io.github.resilience4j.ratelimiter.event.RateLimiterOnFailureEvent;
 import io.github.resilience4j.ratelimiter.event.RateLimiterOnSuccessEvent;
 import io.vavr.collection.Map;
@@ -33,7 +32,7 @@ import static java.lang.System.nanoTime;
 import static java.lang.Thread.currentThread;
 import static java.util.concurrent.locks.LockSupport.parkNanos;
 
-abstract class BaseAtomicLimiter<T > implements RateLimiter {
+abstract class BaseAtomicLimiter<E extends RateLimiterConfig,T extends BaseState<E>> implements RateLimiter {
 
     protected static final long nanoTimeStart = nanoTime();
 
@@ -57,6 +56,18 @@ abstract class BaseAtomicLimiter<T > implements RateLimiter {
 
     protected AtomicReference<T> state() {
         return state;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public boolean acquirePermission(final int permits) {
+        long timeoutInNanos = state().get().getTimeoutInNanos();
+        T modifiedState = updateStateWithBackOff(permits, timeoutInNanos);
+        boolean result = waitForPermissionIfNecessary(timeoutInNanos, modifiedState.getNanosToWait());
+        publishRateLimiterEvent(result, permits);
+        return result;
     }
 
     /**
